requirements.txt
attrdict==2.0.0
numpy
Pillow==6.2.0
# pkg-resources
six==1.11.0
# torch==0.4.0
torchvision==0.2.1
